Introduced gorilla/mux library (routes and handlers)
With this update:
- main.go initialises routes and starts the server
- routes.go contains the routes declaration, accepting now requests
  like /api/tubestatus/{tubeline}
- handlers.go contains all the logic (it could be split into handler
  and service, or controller and service, but maybe it's too early)
- report.go and status.go are 'kind-of' objects. This could be
reverted, I just thought it was a good idea to have a report mapper
in its own file. The mapper is currently returning the same structure
that receives, but in the future the mapping logic can be implemented
just there without chaning the handler, and I think it's nice. It would
be interesting thinking about interfaces, but I'm still reading about
it.